+++ CHANGELOG.md
@@ -6,6 +6,8 @@ this due the [breaking changes](https://github.com/Kotlin/kotlinx.serialization/
 ### Added
 - Recommendation client (#148)
 - Add analytics regions (#170)
+- Natural languages parameter to search query (#173)
+- Logs inner queries (#174)
 
 ### Changed
 - Updated Kotlin to `1.3.70`